Renamed utils parser
